+++ src/components/__tests__/FlawTrackers.spec.ts
@@ -65,9 +65,7 @@ describe('flawTrackers', () => {
     expect(itemsPerPageIndicator.text()).toBe('Per page: 10');
 
     const reduceItemsBtn = subject.find('.trackers-toolbar .tracker-badges .btn .bi-dash-square');
-    for (let i = 0; i < 5; i++) {
-      await reduceItemsBtn.trigger('click');
-    }
+    await reduceItemsBtn.trigger('click');
 
     trackersTableRows = subject.findAll('.affects-trackers .osim-tracker-card table tbody tr');
     rowCount = trackersTableRows.length;
@@ -76,15 +74,37 @@ describe('flawTrackers', () => {
     expect(itemsPerPageIndicator.text()).toBe('Per page: 5');
 
     const increaseItemsBtn = subject.find('.trackers-toolbar .tracker-badges .btn .bi-plus-square');
-    for (let i = 0; i < 5; i++) {
-      await increaseItemsBtn.trigger('click');
-    }
+    await increaseItemsBtn.trigger('click');
 
     trackersTableRows = subject.findAll('.affects-trackers .osim-tracker-card table tbody tr');
     rowCount = trackersTableRows.length;
     expect(rowCount).toBe(6);
     itemsPerPageIndicator = subject.find('.trackers-toolbar .tracker-badges .btn span');
     expect(itemsPerPageIndicator.text()).toBe('Per page: 10');
+  });
+
+  osimFullFlawTest('Show all button works properly', async ({ flaw }) => {
+    const subject = mountFlawTrackers({
+      flaw: flaw as ZodFlawType,
+      relatedFlaws: [flaw as ZodFlawType],
+      displayedTrackers: flaw.affects
+        .flatMap(affect => affect.trackers
+          .map(tracker => ({ ...tracker, ps_module: affect.ps_module })),
+        ),
+      allTrackersCount: 6,
+    });
+
+    let badgeButtons = subject.findAll('.tracker-badges div');
+    const showAllBtn = badgeButtons[1];
+    expect(showAllBtn.text()).toBe('Show All Trackers (6)');
+
+    await showAllBtn.trigger('click');
+
+    const affectsTableRows = subject.findAll('.osim-tracker-card table tbody tr');
+    const rowCount = affectsTableRows.length;
+    expect(rowCount).toBe(6);
+    badgeButtons = subject.findAll('.tracker-badges div');
+    expect(badgeButtons.length).toBe(1);
   });
 
   osimFullFlawTest('Trackers can be sorted by clicking on the date field columns', async ({ flaw }) => {